branches/svl-mts: patch for linux 64-bit compilation
git-svn-id: https://svn.lcsr.jhu.edu/cisst@1605 0959eb1a-6956-4f60-947b-6092a4e22c97

+++ branches/branches/2010-05-05-svl-mts/utilities/multiTaskGlobalComponentManagerFLTK/GCMUITask.cpp
@@ -158,7 +158,7 @@ void callbackPopupSignalSelectionMenu(Fl_Widget * w, void *userdata)
     const Fl_Menu_Item * m = popupMenus->popup(Fl::event_x(), Fl::event_y(), 0, 0, 0);
     if (m) {
         GCMUITask::SignalSelected * signal = new GCMUITask::SignalSelected;
-        signal->Index = reinterpret_cast<int>(m->user_data());
+        signal->Index = (int) reinterpret_cast<size_t>(m->user_data());
         signal->ProcessName = processName;
         signal->ComponentName = componentName;
         signal->InterfaceName = interfaceName;
